:recycle: refactoring: LeftNavMenu 분리
